+++ src/components/NavBar/PeerGroup.tsx
@@ -63,6 +63,11 @@ export default function PeerGroup() {
       <AvatarGroup className={classes.group} max={MAX_PEER_VIEW} onClick={handleViewList}>
         {activePeers.map((peer) => {
           const { username, color, image } = peers[peer.id].presence;
+
+          if (peer.isMine) {
+            return null;
+          }
+
           return (
             <Tooltip key={peer.id} title={peer.isMine ? `[ME] ${username}` : username} data-id={peer.id} arrow>
               <Avatar
